Differentiate between inner APID and bindings, add DidData and Did
